+++ lib/src/inputs.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:meta/meta.dart';
 
-/// Objects that represent an input of the action.
+/// Objects that represent an input of the action
 class Input {
   /// The name of the input,
   /// as written in the YAML file.